move debug repr to not be codecov'd
