+++ src/simpy/__init__.py
@@ -1,21 +1,2 @@
-from .expr import (
-    Rat,
-    acos,
-    asin,
-    atan,
-    cos,
-    cot,
-    csc,
-    debug_repr,
-    diff,
-    e,
-    latex,
-    log,
-    pi,
-    sec,
-    sin,
-    sqrt,
-    symbols,
-    tan,
-)
+from .expr import Rat, acos, asin, atan, cos, cot, csc, diff, e, latex, log, pi, sec, sin, sqrt, symbols, tan
 from .integration import integrate
