+++ src/simpy/debug/__init__.py
@@ -1,0 +1,4 @@
+"""This directory is for debugging utilities. It is not meant to be used in production code.
+
+Not covered by tests, and not guaranteed to be stable or maintained. Use at your own risk.
+"""
