+++ src/simpy/expr.py
@@ -1190,6 +1190,8 @@ class Prod(Associative, Expr):
             if not isinstance(new_prod, Prod):
                 return f"-{_term_repr(new_prod)}"
             if new_prod.is_subtraction:
+                from .debug.utils import debug_repr
+
                 raise ValueError(f"Cannot get repr of {debug_repr(self)}")
             return "-" + new_prod.__repr__()
 
@@ -1318,32 +1320,6 @@ class Prod(Associative, Expr):
 
     def diff(self, var) -> Sum:
         return Sum([Prod([diff(e, var)] + [t for t in self.terms if t is not e]) for e in self.terms])
-
-
-def debug_repr(expr: Expr, *, pedantic: Literal["always", "moderate"] = "moderate") -> str:
-    """Not designed to look pretty; designed to show the structure of the expr.
-
-    pedantic = "always" will show the non-children classes of numbers and symbols
-    """
-    if isinstance(expr, Power):
-        return f"Power({debug_repr(expr.base)}, {debug_repr(expr.exponent)})"
-        # return f'({debug_repr(expr.base)})^{debug_repr(expr.exponent)}'
-    if isinstance(expr, Prod):
-        return "Prod(" + ", ".join([debug_repr(t) for t in expr.terms]) + ")"
-        # return " * ".join([debug_repr(t) for t in expr.terms])
-    if isinstance(expr, Sum):
-        return "Sum(" + ", ".join([debug_repr(t) for t in expr.terms]) + ")"
-        # return " + ".join([debug_repr(t) for t in expr.terms])
-
-    if pedantic == "always":
-        if isinstance(expr, Rat):
-            return f"Rat({expr.value})"
-        if isinstance(expr, Symbol):
-            return f"Symbol({expr.name})"
-        if isinstance(expr, Float):
-            return f"Float({expr.value})"
-
-    return repr(expr)
 
 
 @dataclass
